requirements.txt
pluggy>=0.3,<1.0
PyYAML>=5.3.1
